# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PROOF API
  description: PROOF 서비스 REST API 명세서
  version: 1.0.0
  contact:
    name: PROOF Team
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: 회원가입
      description: 새로운 사용자를 등록합니다
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
                - name
                - email
              properties:
                user_id:
                  type: string
                  description: 사용자 ID (로그인용)
                  example: "john_doe"
                password:
                  type: string
                  description: 비밀번호
                  example: "password123"
                name:
                  type: string
                  description: 이름
                  example: "홍길동"
                email:
                  type: string
                  format: email
                  description: 이메일
                  example: "john@example.com"
                university:
                  type: string
                  description: 대학교 (선택)
                  example: "서울대학교"
                major:
                  type: string
                  description: 전공 (선택)
                  example: "경영학과"
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "회원가입이 완료되었습니다"
                  user_id:
                    type: string
                    example: "john_doe"
        '400':
          description: 잘못된 요청
        '409':
          description: 이미 존재하는 사용자
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: 로그인
      description: 사용자 인증을 수행합니다
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
              properties:
                user_id:
                  type: string
                  description: 사용자 ID
                  example: "john_doe"
                password:
                  type: string
                  description: 비밀번호
                  example: "password123"
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "로그인 성공"
                  user_id:
                    type: string
                    example: "john_doe"
                  name:
                    type: string
                    example: "홍길동"
        '401':
          description: 인증 실패

  # ==================== Users ====================
...
  /api/activities:
    get:
      tags:
        - Activities
      summary: 활동 목록 조회
      description: 추천 활동 목록을 조회합니다
      parameters:
        - name: category
          in: query
          description: 카테고리 필터
          schema:
            type: string
            enum: [공모전, 대외활동, 프로젝트, 기타]
        - name: field
          in: query
          description: 분야 필터
          schema:
            type: string
            enum: [마케팅, 기획, 개발, 디자인, 데이터, 경영, 금융, 기타]
        - name: search
          in: query
          description: 검색어
          schema:
            type: string
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            type: string
            enum: [latest, deadline, popular]
            default: latest
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 활동 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
  /api/activities/{activity_id}:
    get:
      tags:
        - Activities
      summary: 활동 상세 조회
      description: 특정 활동의 상세 정보를 조회합니다
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 활동 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
  /api/activities/recommend:
    get:
      tags:
        - Activities
      summary: 맞춤 활동 추천
      description: 사용자에게 맞춤 활동을 추천합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 추천 활동 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    activity:
                      $ref: '#/components/schemas/Activity'
                    matching_score:
                      type: integer
                      example: 85
                    match_reasons:
                      type: array
                      items:
                        type: string
                      example: ["관심 분야 일치", "역량 개발 기회"]
  /api/activities/{activity_id}/view:
    post:
      tags:
        - Activities
      summary: 조회수 증가
      description: 활동 조회수를 증가시킵니다
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회수 증가 성공

  # ==================== Bookmarks ====================
  /api/bookmarks:
    get:
      tags:
        - Activities
      summary: 북마크 목록 조회
      description: 사용자의 북마크 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: 북마크 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bookmark_id:
                      type: string
                      format: uuid
                    activity:
                      $ref: '#/components/schemas/Activity'
                    created_at:
                      type: string
                      format: date-time

    post:
      tags:
        - Activities
      summary: 북마크 추가
      description: 활동을 북마크에 추가합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - activity_id
              properties:
                activity_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: 북마크 추가 성공
  /api/bookmarks/{bookmark_id}:
    delete:
      tags:
        - Activities
      summary: 북마크 삭제
      description: 북마크를 삭제합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: bookmark_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 북마크 삭제 성공

  # ==================== Job Simulation ====================
...
  /api/survey/questions:
    get:
      tags:
        - Career Survey
      summary: 설문 질문 조회
      description: 커리어 설문 질문을 조회합니다
      parameters:
        - name: survey_type
          in: query
          required: true
          description: 설문 유형
          schema:
            type: string
            enum: [general, spec_check]
      responses:
        '200':
          description: 설문 질문 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  survey_type:
                    type: string
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        question_id:
                          type: integer
                        question_type:
                          type: string
                          enum: [likert, text, single_choice, multiple_choice]
                        question_text:
                          type: string
                        options:
                          type: array
                          items:
                            type: string
  /api/survey/submit:
    post:
      tags:
        - Career Survey
      summary: 일반 설문 제출
      description: 일반 커리어 설문을 제출합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - answers
              properties:
                answers:
                  type: object
                  description: 질문 ID를 키로 하는 답변 객체
                  example:
                    "1": "매우 그렇다"
                    "2": "데이터 분석과 전략 수립"
                    "3": ["협업", "분석력"]
      responses:
        '200':
          description: 설문 제출 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerSurveyResult'
  /api/survey/spec-check/submit:
    post:
      tags:
        - Career Survey
      summary: 스펙 체크 설문 제출
      description: 스펙 체크 설문을 제출합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - job
                - answers
              properties:
                job:
                  type: string
                  description: 선택한 직무
                  example: "마케터"
                answers:
                  type: object
      responses:
        '200':
          description: 설문 제출 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerSurveyResult'
  /api/survey/results:
    get:
      tags:
        - Career Survey
      summary: 설문 결과 목록
      description: 사용자의 설문 결과 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: 설문 결과 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CareerSurveyResult'
  /api/survey/results/{survey_id}:
    get:
      tags:
        - Career Survey
      summary: 설문 결과 상세
      description: 특정 설문 결과를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: survey_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 설문 결과 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareerSurveyResult'
components:
  parameters:
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      description: 사용자 ID
      schema:
        type: string
        example: "john_doe"
  schemas:
...
    Activity:
      type: object
      properties:
        activity_id:
          type: string
          format: uuid
        title:
          type: string
          example: "2024 대학생 마케팅 공모전"
        organization:
          type: string
          example: "한국마케팅협회"
        category:
          type: string
          enum: [공모전, 대외활동, 프로젝트, 기타]
        field:
          type: string
          enum: [마케팅, 기획, 개발, 디자인, 데이터, 경영, 금융, 기타]
        description:
          type: string
        image_url:
          type: string
        external_url:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        benefits:
          type: array
          items:
            type: string
        view_count:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CareerSurveyResult:
      type: object
      properties:
        survey_id:
          type: string
          format: uuid
        user_id:
          type: string
        survey_type:
          type: string
          enum: [general, spec_check]
        answers:
          type: object
          description: 답변 데이터
        preference_top3:
          type: array
          items:
            type: object
            properties:
              job:
                type: string
              score:
                type: integer
          example:
            - job: "마케터"
              score: 85
            - job: "기획자"
              score: 80
            - job: "데이터 분석가"
              score: 75
        fit_top3:
          type: array
          items:
            type: object
            properties:
              job:
                type: string
              score:
                type: integer
        recommended_job:
          type: string
          example: "마케터"
        completed_at:
          type: string
          format: date-time
...